requirements.txt
pandas>=2.1.0
xlrd>=2.0.1
python-dateutil>=2.9.0
pdfplumber>=0.10.3
